Playback speed is now a double between 0 and 2 e.g. 1.03 for +3%.

+++ src/MixPlanner/DomainModel/MixItem.cs
@@ -18,7 +18,7 @@ namespace MixPlanner.DomainModel
 
         public void SetPlaybackSpeed(double value)
         {
-            PlaybackSpeed.SetSpeed((int) value);
+            PlaybackSpeed.SetSpeed(value);
             
         }
     }
